Split the visualization types into a compute half and a React half
One 714-line file held both what a chart is and how it is rendered in the app,
so the settings computation and the echarts models could not be typed without
dragging in redux and the widget components.

The types that describe computation move out: rendering.ts takes the geometry,
the measurement context and the theme, definition.ts takes the chart definition
and the settings definitions, and the two pivoted column shapes join
columns.ts. visualization.ts keeps the types that are React or redux shaped -
VisualizationProps, the pass-through props, VisualizationComponent and
Visualization - and imports the compute names it needs.
