+++ frontend/src/metabase/visualizations/types/index.ts
@@ -1,6 +1,8 @@
 export * from "./click-actions";
 export * from "./columns";
 export * from "./computed-settings";
+export * from "./definition";
 export * from "./hover";
+export * from "./rendering";
 export * from "./visualization";
 export * from "./widget-props";
